src/golang.org/x/net/html/testdata/webkit/tests_innerHTML_1.dat

#data
<body><span>
#errors
#document-fragment
body
#document
| <span>

#data
<span><body>
#errors
#document-fragment
body
#document
| <span>

#data
<span><body>
#errors
#document-fragment
div
#document
| <span>

#data
<body><span>
#errors
#document-fragment
html
#document
| <head>
| <body>
|   <span>

#data
<frameset><span>
#errors
#document-fragment
body
#document
| <span>

#data
<span><frameset>
#errors
#document-fragment
body
#document
| <span>

#data
<span><frameset>
#errors
#document-fragment
div
#document
| <span>

#data
<frameset><span>
#errors
#document-fragment
html
#document
| <head>
| <frameset>

#data
<table><tr>
#errors
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
</table><tr>
#errors
#document-fragment
table
#document
| <tbody>
|   <tr>

#data
<q>
#errors
#document-fragment
table
#document
| <q>

#data
<q>
#errors
#document-fragment
table
#document
| <q>

#data
<q><caption>q
#errors
#document-fragment
table
#document
| <q>
| <caption>
|   "q"

#data
<q><colgroup><col>
#errors
#document-fragment
table
#document
| <q>
| <colgroup>
|   <col>

#data
<q><tbody><tr>
#errors
#document-fragment
table
#document
| <q>
| <tbody>
|   <tr>

#data
<q><tfoot><tr>
#errors
#document-fragment
table
#document
| <q>
| <tfoot>
|   <tr>

#data
<q><thead><tr>
#errors
#document-fragment
table
#document
| <q>
| <thead>
|   <tr>

#data
<q><tr>
#errors
#document-fragment
table
#document
| <q>
| <tbody>
|   <tr>

#data
<q><th>
#errors
#document-fragment
table
#document
| <q>
| <tbody>
|   <tr>
|     <th>

#data
<q><td>
#errors
#document-fragment
table
#document
| <q>
| <tbody>
|   <tr>
|     <td>

#data
<table></table><tbody>
#errors
#document-fragment
caption
#document
| <table>

#data
</table><span>
#errors
#document-fragment
caption
#document
| <span>

#data
<span></table>
#errors
#document-fragment
caption
#document
| <span>

#data
</caption><span>
#errors
#document-fragment
caption
#document
| <span>

#data
<span></caption><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><caption><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><col><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><colgroup><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><html><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tbody><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><td><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tfoot><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><thead><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><th><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span><tr><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
<span></table><span>
#errors
#document-fragment
caption
#document
| <span>
|   <span>

#data
</colgroup><col>
#errors
#document-fragment
colgroup
#document
| <col>

#data
<q><col>
#errors
#document-fragment
colgroup
#document
| <col>

#data
<caption><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
<col><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
<colgroup><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
<tbody><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
<tfoot><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
<thead><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
</table><q>
#errors
#document-fragment
tbody
#document
| <q>

#data
<q><tr>
#errors
#document-fragment
tbody
#document
| <q>
| <tr>

#data
<q><td>
#errors
#document-fragment
tbody
#document
| <q>
| <tr>
|   <td>

#data
<q><td>
#errors
#document-fragment
tbody
#document
| <q>
| <tr>
|   <td>

#data
<q><td>
#errors
#document-fragment
tbody
#document
| <q>
| <tr>
|   <td>

#data
<td><table><tbody><q><tr>
#errors
#document-fragment
tbody
#document
| <tr>
|   <td>
|     <q>
|     <table>
|       <tbody>
|         <tr>

#data
</tr><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<td><table><q><tr></tr><tr>
#errors
#document-fragment
tr
#document
| <td>
|   <q>
|   <table>
|     <tbody>
|       <tr>
|       <tr>

#data
<caption><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<col><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<colgroup><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<tbody><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<tfoot><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<thead><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<tr><td>
#errors
#document-fragment
tr
#document
| <td>

#data
</table><td>
#errors
#document-fragment
tr
#document
| <td>

#data
<td><table></table><td>
#errors
#document-fragment
tr
#document
| <td>
|   <table>
| <td>

#data
<td><table></table><td>
#errors
#document-fragment
tr
#document
| <td>
|   <table>
| <td>

#data
<caption><q>
#errors
#document-fragment
td
#document
| <q>

#data
<col><q>
#errors
#document-fragment
td
#document
| <q>

#data
<colgroup><q>
#errors
#document-fragment
td
#document
| <q>

#data
<tbody><q>
#errors
#document-fragment
td
#document
| <q>

#data
<tfoot><q>
#errors
#document-fragment
td
#document
| <q>

#data
<th><q>
#errors
#document-fragment
td
#document
| <q>

#data
<thead><q>
#errors
#document-fragment
td
#document
| <q>

#data
<tr><q>
#errors
#document-fragment
td
#document
| <q>

#data
</table><q>
#errors
#document-fragment
td
#document
| <q>

#data
</tbody><q>
#errors
#document-fragment
td
#document
| <q>

#data
</td><q>
#errors
#document-fragment
td
#document
| <q>

#data
</tfoot><q>
#errors
#document-fragment
td
#document
| <q>

#data
</thead><q>
#errors
#document-fragment
td
#document
| <q>

#data
</th><q>
#errors
#document-fragment
td
#document
| <q>

#data
</tr><q>
#errors
#document-fragment
td
#document
| <q>

#data
<table><td><td>
#errors
#document-fragment
td
#document
| <table>
|   <tbody>
|     <tr>
|       <td>
|       <td>

#data
</select><option>
#errors
#document-fragment
select
#document
| <option>

#data
<input><option>
#errors
#document-fragment
select
#document
| <option>

#data
<keygen><option>
#errors
#document-fragment
select
#document
| <option>

#data
<textarea><option>
#errors
#document-fragment
select
#document
| <option>

#data
</html><!--abc-->
#errors
#document-fragment
html
#document
| <head>
| <body>
| <!-- abc -->

#data
</frameset><frame>
#errors
#document-fragment
frameset
#document
| <frame>

#data
#errors
#document-fragment
html
#document
| <head>
| <body>
